Add support for using expirations of more than 30 days
In Memcached the expiration times are specified in seconds
as unsigned integer value. The expiration can take following
values:

1) `0` - never expire
2) `Up to 30 days` i.e. 60x60x24x30 = 2.592.000 seconds
3) Any value `greater than 30 days` is intepreted as Unix timestamp e.g.
can be used to expire cache on specific date.

In order to support expiration values of more than 30 days, the
check is performed when obtaining configured expiration when cache
client `set()` and `touch()` calls. If the given expiration in seconds
is greater than 30 days, we add the expiration value to the current
instant of UTC timezone (Unix timestamp). In case expiration is less than
or equals to 30 days, we do not perform any recalculation and use the
existing expiration for cache client's `set()` and `touch()` calls.

In this way we can continue using the `memcached.cache.expiration` &
`memcached.cache.expiration-per-cache` configuration properties as
time-to-live value i.e. cache entry expiration. This change will not
introduce any breaking change (e.g. expiring cache on specific date),
being we did not allow expiration values of more than 30 days.

Resolves #152